requirements.txt
pyyaml
dotmap
fs
fs-s3fs
fs.sshfs
python-baseconv
regex
observable
gitignore_parser
orator
click